requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
pymysql>=1.1.0
python-dotenv>=1.0.0
lxml>=5.1.0
